dockerize-applications/python asgi/requirements.txt
aioredis==1.1.0
locustio==0.9.0
uvicorn==0.3.5
